requirements.txt
llama-index>=0.9.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
gradio>=3.40.0
pypdf>=3.12.0
faiss-cpu>=1.7.4
